refactor rewards claiming
