+++ novawallet/Modules/Staking/GeneralStaking/ClaimRewards/StakingClaimRewardsProtocols.swift
@@ -1,16 +1,22 @@
 import Foundation
 
-protocol StakingClaimRewardsViewProtocol: SCLoadableControllerProtocol {
+protocol StakingGenericRewardsViewProtocol: SCLoadableControllerProtocol {
     func didReceiveAmount(viewModel: BalanceViewModelProtocol)
     func didReceiveWallet(viewModel: DisplayWalletViewModel)
     func didReceiveAccount(viewModel: DisplayAddressViewModel)
     func didReceiveFee(viewModel: BalanceViewModelProtocol?)
-    func didReceiveClaimStrategy(viewModel: StakingClaimRewardsStrategy)
 }
 
-protocol StakingClaimRewardsPresenterProtocol: AnyObject {
+protocol StakingGenericRewardsPresenterProtocol: AnyObject {
     func setup()
     func confirm()
     func selectAccount()
+}
+
+protocol StakingClaimRewardsViewProtocol: StakingGenericRewardsViewProtocol {
+    func didReceiveClaimStrategy(viewModel: StakingClaimRewardsStrategy)
+}
+
+protocol StakingClaimRewardsPresenterProtocol: StakingGenericRewardsPresenterProtocol {
     func toggleClaimStrategy()
 }
