+++ novawallet/Modules/Staking/GeneralStaking/ClaimRewards/StakingClaimRewardsViewController.swift
@@ -1,77 +1,33 @@
 import UIKit
 import SoraFoundation
 
-final class StakingClaimRewardsViewController: UIViewController, ViewHolder {
-    typealias RootViewType = StakingClaimRewardsViewLayout
-
-    let presenter: StakingClaimRewardsPresenterProtocol
+final class StakingClaimRewardsViewController: StakingGenericRewardsViewController<StakingClaimRewardsViewLayout> {
+    var presenter: StakingClaimRewardsPresenterProtocol? {
+        basePresenter as? StakingClaimRewardsPresenterProtocol
+    }
 
     init(presenter: StakingClaimRewardsPresenterProtocol, localizationManager: LocalizationManagerProtocol) {
-        self.presenter = presenter
-
-        super.init(nibName: nil, bundle: nil)
-
-        self.localizationManager = localizationManager
+        super.init(basePresenter: presenter, localizationManager: localizationManager)
     }
 
-    @available(*, unavailable)
-    required init?(coder _: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    override func loadView() {
-        view = StakingClaimRewardsViewLayout()
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
+    override func onViewDidLoad() {
+        super.onViewDidLoad()
 
         setupHandlers()
         setupLocalization()
-
-        presenter.setup()
     }
 
     private func setupLocalization() {
-        let languages = selectedLocale.rLanguages
-
-        title = R.string.localizable.stakingClaimRewards(preferredLanguages: languages)
-
-        rootView.actionButton.imageWithTitleView?.title = R.string.localizable
-            .commonConfirm(preferredLanguages: selectedLocale.rLanguages)
-
-        rootView.walletCell.titleLabel.text = R.string.localizable.commonWallet(
-            preferredLanguages: selectedLocale.rLanguages
-        )
-
-        rootView.accountCell.titleLabel.text = R.string.localizable.commonAccount(
-            preferredLanguages: selectedLocale.rLanguages
-        )
-
-        rootView.networkFeeCell.rowContentView.locale = selectedLocale
-
         rootView.restakeCell.titleLabel.text = R.string.localizable.stakingRestakeTitle_v2_2_0(
-            preferredLanguages: languages
+            preferredLanguages: selectedLocale.rLanguages
         )
 
         rootView.restakeCell.subtitleLabel.text = R.string.localizable.stakingRestakeMessage(
-            preferredLanguages: languages
+            preferredLanguages: selectedLocale.rLanguages
         )
     }
 
     private func setupHandlers() {
-        rootView.actionButton.addTarget(
-            self,
-            action: #selector(actionConfirm),
-            for: .touchUpInside
-        )
-
-        rootView.accountCell.addTarget(
-            self,
-            action: #selector(actionSelectAccount),
-            for: .touchUpInside
-        )
-
         rootView.restakeCell.switchControl.addTarget(
             self,
             action: #selector(actionToggleClaimStrategy),
@@ -79,46 +35,14 @@ final class StakingClaimRewardsViewController: UIViewController, ViewHolder {
         )
     }
 
-    @objc private func actionConfirm() {
-        presenter.confirm()
-    }
-
-    @objc private func actionSelectAccount() {
-        presenter.selectAccount()
-    }
-
     @objc private func actionToggleClaimStrategy() {
-        presenter.toggleClaimStrategy()
+        presenter?.toggleClaimStrategy()
     }
 }
 
 extension StakingClaimRewardsViewController: StakingClaimRewardsViewProtocol {
-    func didReceiveAmount(viewModel: BalanceViewModelProtocol) {
-        rootView.amountView.bind(viewModel: viewModel)
-    }
-
-    func didReceiveWallet(viewModel: DisplayWalletViewModel) {
-        rootView.walletCell.bind(viewModel: viewModel.cellViewModel)
-    }
-
-    func didReceiveAccount(viewModel: DisplayAddressViewModel) {
-        rootView.accountCell.bind(viewModel: viewModel.cellViewModel)
-    }
-
-    func didReceiveFee(viewModel: BalanceViewModelProtocol?) {
-        rootView.networkFeeCell.rowContentView.bind(viewModel: viewModel)
-    }
-
     func didReceiveClaimStrategy(viewModel: StakingClaimRewardsStrategy) {
         let shouldRestake = viewModel == .restake
         rootView.restakeCell.switchControl.setOn(shouldRestake, animated: false)
-    }
-}
-
-extension StakingClaimRewardsViewController: Localizable {
-    func applyLocalization() {
-        if isViewLoaded {
-            setupLocalization()
-        }
     }
 }
